third_party/WebKit/LayoutTests/fast/mediastream/RTCPeerConnection-ice-expected.txt

Tests the RTCPeerConnection Ice functionality.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc = new webkitRTCPeerConnection(null, null); did not throw exception.
PASS iceConnectionState is completed
PASS pc.addIceCandidate(null, addIceCandidateSuccess, addIceCandidateFailure); threw exception TypeError: Failed to execute 'addIceCandidate' on 'RTCPeerConnection': parameter 1 is not of type 'RTCIceCandidate'..
PASS pc.addIceCandidate(iceCandidate, null, addIceCandidateFailure); threw exception TypeError: Failed to execute 'addIceCandidate' on 'RTCPeerConnection': The callback provided as parameter 2 is not a function..
PASS pc.addIceCandidate(iceCandidate, addIceCandidateSuccess, null); threw exception TypeError: Failed to execute 'addIceCandidate' on 'RTCPeerConnection': The callback provided as parameter 3 is not a function..
PASS pc.addIceCandidate(iceCandidate, addIceCandidateSuccess, addIceCandidateFailure); did not throw exception.
PASS addIceCandidateSuccess was called.
PASS iceConnectionState is closed.
PASS successfullyParsed is true

TEST COMPLETE
